Create User::FeedMethods concern
